# docs/swagger.yaml
basePath: /
definitions:
...
  dto.CourseScheduleResponse:
    properties:
      course:
        type: integer
      generated_at:
        type: string
      groups:
        items:
          $ref: '#/definitions/dto.ScheduleGroupSummaryDTO'
        type: array
      weeks:
        items:
          $ref: '#/definitions/dto.WeekScheduleResponse'
        type: array
    type: object
...
  dto.PlanItemResponse:
    properties:
      course:
        type: integer
      planned_pairs:
        type: integer
      subject:
        type: string
    type: object
  dto.PlanUpsertRequest:
    properties:
      course:
        type: integer
      planned_pairs:
        type: integer
      subject:
        type: string
    type: object
...
  dto.ScheduleGroupRequest:
    properties:
      department:
        type: string
      id:
        type: string
      name:
        type: string
      specialty:
        type: string
    type: object
  dto.ScheduleGroupSummaryDTO:
    properties:
      count:
        type: integer
      id:
        type: string
      name:
        type: string
    type: object
  dto.ScheduleLessonRequest:
    properties:
      cancelled:
        type: boolean
      comment:
        type: string
      date:
        type: string
      day:
        type: string
      day_number:
        type: integer
      duration:
        type: integer
      frequency:
        type: string
      google_sheet_id:
        type: string
      group:
        type: string
      pair:
        type: integer
      period_end:
        type: string
      period_start:
        type: string
      room:
        type: string
      subgroup:
        type: string
      subject:
        type: string
      teacher:
        type: string
      time:
        type: string
      type:
        type: string
    type: object
  dto.ScheduleLessonResponse:
    properties:
      cancelled:
        type: boolean
      comment:
        type: string
      date:
        type: string
      day:
        type: string
      day_number:
        type: integer
      duration:
        type: integer
      frequency:
        type: string
      google_sheet_id:
        type: string
      group:
        type: string
      pair:
        type: integer
      period_end:
        type: string
      period_start:
        type: string
      room:
        type: string
      subgroup:
        type: string
      subject:
        type: string
      teacher:
        type: string
      time:
        type: string
      type:
        type: string
      week_number:
        type: integer
    type: object
...
  dto.WeekScheduleRequest:
    properties:
      course:
        type: integer
      date_range:
        type: string
      generated_at:
        type: string
      groups:
        items:
          $ref: '#/definitions/dto.ScheduleGroupRequest'
        type: array
      lessons:
        items:
          $ref: '#/definitions/dto.ScheduleLessonRequest'
        type: array
      name:
        type: string
      semester:
        type: integer
      week_number:
        type: integer
    type: object
  dto.WeekScheduleResponse:
    properties:
      course:
        type: integer
      date_range:
        type: string
      generated_at:
        type: string
      groups:
        items:
          $ref: '#/definitions/dto.ScheduleGroupSummaryDTO'
        type: array
      lessons:
        items:
          $ref: '#/definitions/dto.ScheduleLessonResponse'
        type: array
      name:
        type: string
      semester:
        type: integer
      week_number:
        type: integer
    type: object
...
  http.ErrorResponse:
    properties:
      error:
        example: invalid request body
        type: string
      message:
        example: course query parameter is required
        type: string
      status:
        example: error
        type: string
    type: object
...
  http.PlanUpsertSuccessResponse:
    properties:
      status:
        example: success
        type: string
    type: object
...
  http.ScheduleImportSuccessResponse:
    properties:
      course:
        example: 3
        type: integer
      lessons_count:
        example: 120
        type: integer
      ok:
        example: true
        type: boolean
      status:
        example: success
        type: string
      updated_at:
        example: "2026-05-17T19:00:00Z"
        type: string
      week_number:
        example: 14
        type: integer
    type: object
...
info:
  contact: {}
  description: API for schedule import/read, attendance processing and study plan
    management.
  title: CourseJob API
  version: "1.0"
paths:
...
  /api/v1/plan:
    get:
      description: Returns planned pairs grouped by subject for the selected course.
      parameters:
      - description: Course number (1-4)
        in: query
        name: course
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/dto.PlanItemResponse'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/http.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/http.ErrorResponse'
      summary: Get plan by course
      tags:
      - Plan
    put:
      consumes:
      - application/json
      description: Creates or updates planned pairs by course and subject.
      parameters:
      - description: Plan upsert payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.PlanUpsertRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/http.PlanUpsertSuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/http.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/http.ErrorResponse'
      summary: Upsert plan item
      tags:
      - Plan
...
  /api/v1/schedule:
    get:
      description: Returns schedule grouped by weeks with optional filters.
      parameters:
      - description: Course number (1-4)
        in: query
        name: course
        required: true
        type: integer
      - description: Week number
        in: query
        name: week
        type: integer
      - description: Group filter
        in: query
        name: group
        type: string
      - description: Day filter
        in: query
        name: day
        type: string
      - description: Lesson type filter
        in: query
        name: type
        type: string
      - description: Teacher filter
        in: query
        name: teacher
        type: string
      - description: Subject filter
        in: query
        name: subject
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.CourseScheduleResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/http.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/http.ErrorResponse'
      summary: Get schedule by course
      tags:
      - Schedule
    post:
      consumes:
      - application/json
      description: Imports one course schedule chunk and publishes an SSE update event.
      parameters:
      - description: Week schedule payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.WeekScheduleRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/http.ScheduleImportSuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/http.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/http.ErrorResponse'
      summary: Import schedule
      tags:
      - Schedule
  /api/v1/sse/schedule:
    get:
      description: Opens an SSE stream and sends `schedule_updated` events when schedule
        import succeeds.
      produces:
      - text/event-stream
      responses:
        "200":
          description: SSE stream
          schema:
            type: string
        "405":
          description: Method Not Allowed
          schema:
            $ref: '#/definitions/http.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/http.ErrorResponse'
      summary: Subscribe to schedule updates
      tags:
      - Schedule
...
schemes:
- https
swagger: "2.0"